broke out firebase imports and inits to seperate file
